removed some unused components, moved task reducer to workflows
