+++ server/modules/pool.js
@@ -21,8 +21,8 @@ if (process.env.DATABASE_URL) {
   };
 } else {
   config = {
-    // user: null || 'con',
-    // password: null || 'secretpass',
+    user: null || 'con',
+    password: null || 'secretpass',
     host: 'localhost', // Server hosting the postgres database
     port: 5432, // env var: PGPORT
     database: 'svenplan_db',
