--- server/routes/haleytask.router.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const pool = require('../modules/pool');
-const router = express.Router();
-
-
-
-module.exports = router;
