+++ src/components/WorkflowEditor/AddTaskHaley/subcomponents/AddTask2.js
@@ -46,7 +46,7 @@ class AddTask2 extends Component {
 
 const putReduxStateOnProps = (reduxState) => ({
     user: reduxState.user,
-    task: reduxState.task
+    task: reduxState.workflow
 });
 
 export default connect(putReduxStateOnProps)(withRouter(AddTask2));
