+++ src/components/WorkflowEditor/AddTaskHaley/subcomponents/AddTask3.js
@@ -77,7 +77,7 @@ class AddTask3 extends Component {
 
 const putReduxStateOnProps = (reduxState) => ({
     user: reduxState.user,
-    task: reduxState.task
+    task: reduxState.workflow
 });
 
 export default connect(putReduxStateOnProps)(withRouter(AddTask3));
